+++ src/main/java/com/example/spot/member/presentation/controller/legacy/MemberController.java
@@ -33,6 +33,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import static com.example.spot.auth.presentation.dto.google.GoogleExampleResponse.EXAMPLE_RESPONSE;
 
+// TODO : 이 클래스는 구글 로그인 API를 위한 레거시 컨트롤러입니다. 추후 리팩토링 예정입니다.
+
 @Deprecated
 @RestController
 @RequestMapping("/spot")
@@ -41,37 +43,6 @@ import static com.example.spot.auth.presentation.dto.google.GoogleExampleRespons
 public class MemberController {
 
     private final MemberService memberService;
-    @Tag(name = "테스트 용 API", description = "테스트 용 API")
-    @Operation(summary = "!테스트 용! [회원 생성] 테스트 용 회원 생성 API",
-        description = """
-            ## [테스트 용 회원 생성] 임의의 정보를 가진 회원 객체가 생성 됩니다. 
-            다른 API를 테스트 하기 위해 회원이 필요한 경우 사용해주세요.
-            회원의 관심 분야 및 지역을 입력 받습니다.   
-           생성된 회원의 ID와 Email이 반환 됩니다. """)
-    @PostMapping("/members/test")
-    public ApiResponse<MemberResponseDTO.MemberTestDTO> testMember(
-        @RequestBody @Valid MemberRequestDTO.MemberInfoListDTO memberInfoListDTO){
-        MemberTestDTO dto = memberService.testMember(memberInfoListDTO);
-        return ApiResponse.onSuccess(SuccessStatus._MEMBER_CREATED, dto);
-    }
-
-    @Tag(name = "테스트 용 API", description = "테스트 용 API")
-    @Operation(summary = "!테스트 용! [회원 권한 부여] 관리자 권한 부여 API",
-        description = """
-            ## [회원 권한 부여] 해당하는 회원에게 관리자 권한을 부여합니다.
-            테스트를 위해 구현한 테스트 용 API입니다.
-            회원의 ID를 입력 받아 관리자 권한을 부여합니다.
-            성공 여부와 회원 ID가 반환 됩니다.
-             """)
-    @PostMapping("/members/test/admin")
-    public ApiResponse<MemberResponseDTO.MemberUpdateDTO> toAdmin(){
-        MemberUpdateDTO dto = memberService.toAdmin(SecurityUtils.getCurrentUserId());
-        return ApiResponse.onSuccess(SuccessStatus._MEMBER_CREATED, dto);
-    }
-
-
-
-
 
     @Tag(name = "구글 로그인 API", description = "구글 OAuth2 로그인 API")
     @Operation(summary = "[구글 로그인] 구글 로그인/회원가입 API",
